requirements.txt
black
pyright
jupyter
python-dotenv
openai
fastapi[all]
